# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/spaces/{space_id}/teams/": {
      "get": {
        "summary": "Get Teams",
        "operationId": "get_teams_spaces__space_id__teams__get",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamDto"
                  },
                  "title": "Response Get Teams Spaces  Space Id  Teams  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Teams",
        "operationId": "create_teams_spaces__space_id__teams__post",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamsImportRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportedTeamResultDto"
                  },
                  "title": "Response Create Teams Spaces  Space Id  Teams  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove Teams",
        "operationId": "remove_teams_spaces__space_id__teams__delete",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamsDeleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamDeletedDto"
                  },
                  "title": "Response Remove Teams Spaces  Space Id  Teams  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/{space_id}/teams/{team_id}": {
      "get": {
        "summary": "Get Team Information",
        "operationId": "get_team_information_spaces__space_id__teams__team_id__get",
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          },
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamInfoDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/spaces/{space_id}/user_room/team": {
      "delete": {
        "summary": "Remove Users From Team",
        "operationId": "remove_users_from_team_spaces__space_id__user_room_team_delete",
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Space Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsersAndTeamsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsersAndTeamResDto"
                  },
                  "title": "Response Remove Users From Team Spaces  Space Id  User Room Team Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteTeamDto": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "rid": {
            "type": "string",
            "title": "Rid"
          }
        },
        "type": "object",
        "required": [
          "id",
          "rid"
        ],
        "title": "DeleteTeamDto"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportedTeamResultDto": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/TeamCreateDto"
          },
          "created_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "request"
        ],
        "title": "ImportedTeamResultDto"
      },
...
      "RoomDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "broadcast": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Broadcast"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "t": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "T"
          },
          "msgs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msgs"
          },
          "usersCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Userscount"
          },
          "u": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShortUserDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "ro": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ro"
          },
          "default": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "announcement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Announcement"
          }
        },
        "type": "object",
        "required": [
          "_id"
        ],
        "title": "RoomDto"
      },
...
      "RoomUserDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "username",
          "name",
          "status"
        ],
        "title": "RoomUserDto"
      },
...
      "ShortUserDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "_id",
          "username"
        ],
        "title": "ShortUserDto"
      },
...
      "TeamCreateDto": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "team_type": {
            "type": "integer",
            "title": "Team Type"
          },
          "disable_system_messages": {
            "type": "boolean",
            "title": "Disable System Messages",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "team_type"
        ],
        "title": "TeamCreateDto"
      },
      "TeamDeletedDto": {
        "properties": {
          "team": {
            "type": "string",
            "title": "Team"
          },
          "rooms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rooms",
            "default": []
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "team",
          "success"
        ],
        "title": "TeamDeletedDto"
      },
      "TeamDto": {
        "properties": {
          "_id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "createdBy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShortUserDto"
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "roomId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roomid"
          }
        },
        "type": "object",
        "required": [
          "_id"
        ],
        "title": "TeamDto"
      },
      "TeamInfoDto": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/RoomUserDto"
            },
            "type": "array",
            "title": "Users"
          },
          "rooms": {
            "items": {
              "$ref": "#/components/schemas/RoomDto"
            },
            "type": "array",
            "title": "Rooms"
          }
        },
        "type": "object",
        "required": [
          "users",
          "rooms"
        ],
        "title": "TeamInfoDto"
      },
...
      "TeamsDeleteDto": {
        "properties": {
          "teams": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Teams"
          },
          "delete_linked_rooms": {
            "type": "boolean",
            "title": "Delete Linked Rooms",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "teams"
        ],
        "title": "TeamsDeleteDto"
      },
      "TeamsImportRequestDto": {
        "properties": {
          "teams": {
            "items": {
              "$ref": "#/components/schemas/TeamCreateDto"
            },
            "type": "array",
            "title": "Teams"
          }
        },
        "type": "object",
        "required": [
          "teams"
        ],
        "title": "TeamsImportRequestDto"
      },
...
      "UsersAndTeamResDto": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "user_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "User List"
          },
          "team": {
            "type": "string",
            "title": "Team"
          }
        },
        "type": "object",
        "required": [
          "user_list",
          "team"
        ],
        "title": "UsersAndTeamResDto"
      },
      "UsersAndTeamsDto": {
        "properties": {
          "users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Users"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/DeleteTeamDto"
            },
            "type": "array",
            "title": "Teams"
          },
          "ban_in_rooms": {
            "type": "boolean",
            "title": "Ban In Rooms",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "users",
          "teams"
        ],
        "title": "UsersAndTeamsDto"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}